+++ crypto-scraper/app/Console/Base/Common/KafkaConProducer.php
@@ -1,8 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Kafka;
+namespace App\Console\Base\Common;
 
+use App\Kafka\ConProducerFeatures;
 use Illuminate\Console\Command;
 
 abstract class KafkaConProducer extends Command {
