+++ crypto-scraper/app/Console/Base/Common/KafkaProducer.php
@@ -1,8 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Kafka;
+namespace App\Console\Base\Common;
 
+use App\Kafka\ProducerFeatures;
 use Illuminate\Console\Command;
 use RdKafka\Producer;
 use RdKafka\ProducerTopic;
