+++ crypto-scraper/app/Console/Commands/Common/FetchWalletExplorer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Console\Commands;
+namespace App\Console\Commands\Common;
 
 use App\Console\Base\Utils;
 use App\Models\Pg\Category;
