+++ crypto-scraper/app/Console/Commands/Common/ReadLocalStorage.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Console\Commands;
+namespace App\Console\Commands\Common;
 
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Storage;
