+++ crypto-scraper/app/Console/Commands/Tests/KafkaContinuous/TestKafkaConsumer.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Console\Commands\Tests\KafkaContinuous;
 
-use App\Kafka\KafkaConsumer;
+use App\Console\Base\Common\KafkaConsumer;
 use RdKafka\Message;
 
 class TestKafkaConsumer extends KafkaConsumer {
