+++ crypto-scraper/app/Console/Commands/Tests/KafkaSingle/TestKafkaProducer.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Console\Commands\Tests\KafkaSingle;
 
-use App\Kafka\KafkaProducer;
+use App\Console\Base\Common\KafkaProducer;
 
 class TestKafkaProducer extends KafkaProducer {
